requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
requests==2.31.0
redis==5.0.1
Pillow==10.2.0